pyproject.toml
[project]
name = "object-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "efficientnet-pytorch>=0.7.1",
    "joblib>=1.5.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pytorch-lightning>=2.5.6",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "ultralytics>=8.3.231",
]
